requirements.txt
PyYAML==3.11
docopt==0.6.2
evernote==1.25.0
google-api-python-client==1.3.1
httplib2==0.9
oauth2==1.5.211
oauth2client==1.3
pyaml==14.05.7
pyasn1==0.1.7
pyasn1-modules==0.0.5
rsa==3.1.4
simplejson==3.6.4
uritemplate==0.6
wsgiref==0.1.2
